toolchain: make emar support .lo files
When using alwayslink in cc_library, bazel will produce .lo files and
pass --whole-archive to the linker. While this is fine in theory,
emar.py will add fullpath hashes to the filename in the archive to
prevent basenames collisions. However, it only does it if the archive
has a .a extension. This commit makes it so that the output name passed
to emar.py is a .a extension, which is then moved to the original .lo
file.

Signed-off-by: Steeve Morin <steeve@zen.ly>